requirements.txt
# Resolve packages from PyPI only (avoids mixed corporate index configs).
--index-url https://pypi.org/simple
--trusted-host pypi.org

google-genai>=1.0.0
pypdf>=4.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
